+++ src/Helpers/Stack.php
@@ -23,13 +23,13 @@ class Stack
         return \array_pop($this->array);
     }
 
-    public function push(&$item)
+    public function push($item)
     {
         return \array_push($this->array, $item);
     }
 
-    public function getAll()
+    public function in($element)
     {
-        return $this->array;
+        return \in_array($element, $this->array, true);
     }
 }