+++ src/Subdumpers/SubdumperInterface.php
@@ -16,7 +16,7 @@ namespace Awesomite\VarDumper\Subdumpers;
  */
 interface SubdumperInterface
 {
-    public function supports(&$var);
+    public function supports($var);
 
-    public function dump(&$var);
+    public function dump($var);
 }