+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.10.0 (????-??-??)
+
+* Convert `\t`, `\r`, `\0`, `\x0B` to visible characters
+* Split multiline text by white character whenever it is possible e.g.:
+```
+string(768)
+    › Lorem ipsum dolor sit amet, consectetur adipiscing
+    ›  elit. Proin nibh augue, suscipit a, scelerisque
+    › sed, lacinia in, mi. Cras vel lorem. Etiam
+    › pellentesque aliquet tellus. Phasellus pharetra
+    › nulla ac diam. Quisque semper justo at risus.
+    › Donec venenatis, turpis vel hendrerit interdum,
+    › dui ligula ultricies purus, sed posuere libero dui
+    ›  id orci. Nam congue, pede vitae dapibus aliquet,
+    › elit magna vulpu...
+```
+
 ## 0.9.0 (2017-12-21)
 
 * Class `Awesomite\VarDumper\LightVarDumper` becomes `final`